REF: Rename classifier modules to _lowercase.py convention
