+++ skordinal/classifiers/__init__.py
@@ -1,10 +1,10 @@
 """Ordinal classification classifiers module."""
 
-from .NNOP import NNOP
-from .NNPOM import NNPOM
-from .OrdinalDecomposition import OrdinalDecomposition
-from .REDSVM import REDSVM
-from .SVOREX import SVOREX
+from ._nnop import NNOP
+from ._nnpom import NNPOM
+from ._ordinal_decomposition import OrdinalDecomposition
+from ._redsvm import REDSVM
+from ._svorex import SVOREX
 
 __all__ = [
     "NNOP",
