+++ skordinal/classifiers/_redsvm.py
