+++ skordinal/classifiers/tests/test_ordinal_decomposition.py
@@ -4,7 +4,7 @@ import numpy as np
 import numpy.testing as npt
 import pytest
 
-from skordinal.classifiers.OrdinalDecomposition import OrdinalDecomposition
+from skordinal.classifiers import OrdinalDecomposition
 
 
 @pytest.fixture
